moves TC syntheticResolution to the ide-plugin
