+++ idea-plugin/src/main/kotlin/arrow/meta/plugin/idea/IdeMetaPlugin.kt
@@ -10,6 +10,7 @@ import arrow.meta.plugin.idea.plugins.higherkinds.higherKindsIdePlugin
 import arrow.meta.plugin.idea.plugins.initial.initialIdeSetUp
 import arrow.meta.plugin.idea.plugins.nothing.nothingIdePlugin
 import arrow.meta.plugin.idea.plugins.optics.opticsIdePlugin
+import arrow.meta.plugin.idea.plugins.typeclasses.typeclassesIdePlugin
 import kotlin.contracts.ExperimentalContracts
 
 class IdeMetaPlugin : MetaPlugin(), IdeInternalRegistry, IdeSyntax {
@@ -17,8 +18,9 @@ class IdeMetaPlugin : MetaPlugin(), IdeInternalRegistry, IdeSyntax {
   override fun intercept(ctx: CompilerContext): List<Plugin> =
     super.intercept(ctx) +
       initialIdeSetUp +
-      nothingIdePlugin +
+      higherKindsIdePlugin +
+      typeclassesIdePlugin +
       comprehensionsIdePlugin +
       opticsIdePlugin +
-      higherKindsIdePlugin
+      nothingIdePlugin
 }
